Improve the UI switching screens
Make sure to reset collections when switching screens
Make sure the user id is correctly set when switching users
Make sure the Upload and Add directory button are not disabled for regular users.
Make sure the Users query is behaving the right way by initing a user meta if not available.