Prepare Maintenance Release 1.1.2
Improve the Gutenberg Block making sure the MediaThèque icon is inline with the other Blocks icons and adding this icon as a placeholder to the Add User Media button.
Stop checking the Error status when using assertErrorResponse() in tests as WordPress changed one of the error code it in 5.0 and it would make previous versions tests fail.