Make sure a directory still exists on the server before deleting it
The recent changes to WordPress Develop (src) needing it to be built to test it introduced an interesting case where a directory could exists into the database but no more on the server. In this case instead of returning false too early when a user tries to delete it from the UI, we need to let the function carry on so that the directory can be deleted from the database.