Add a fallback to wpApiSettings in case Gutenberg is active.
As Gutenberg is *arbitrary* deregistering the `wp-api-request` script to replace it with a new implementation, we cannot trust anymore WordPress Core way of setting the wpApiSettings. As a result to be sure to be able to get the REST nonce for this plugin need, we have to add it to our script in case Gutenberg is active.

See https://github.com/WordPress/gutenberg/pull/7329